All done except Test1.java and Test2.java
I cleaned up all of the DAO's and their respective data objects. Now all
we need to do is finish Test1.java with some queries and change the data
once or twice. Then we just need to write Test2.java to handle getting
everything at once to make sure we did in fact save to a database and
not just the RAM.